refactor(frontend): extract useVersionMutations composable (W7)
Move version rollback/delete + per-version publish-records dialog (rollback/
delete loading refs, dialog state, versionPublishRecordDialogTitle computed,
rollbackToVersion/deleteVersion/openVersionPublishRecords) into
composables/useVersionMutations.js, smoke-gated. workflow/versionById/
publishRecordsByVersionId injected; the W3 disabled-reason wrappers,
loadWorkflowDetail and backToPublishRecords passed as lazy forward refs.

Verified: WorkflowDetail mount smoke + full suite (27 files / 116 passed);
npm run build, npm run lint (0 errors). WorkflowDetail.vue 2131 -> 2047.
