+++ frontend/src/views/workflows/WorkflowDetail.vue
@@ -867,6 +867,7 @@ import {
 import { useInlineFieldEditing } from './composables/useInlineFieldEditing'
 import { useScheduleForm } from './composables/useScheduleForm'
 import { useVersionCompare } from './composables/useVersionCompare'
+import { useVersionMutations } from './composables/useVersionMutations'
 
 const route = useRoute()
 const router = useRouter()
@@ -881,11 +882,6 @@ const actionLoading = reactive({})
 const backfillDialogVisible = ref(false)
 const backfillTarget = ref(null)
 const changeMode = ref('list')
-const rollbackLoadingVersionId = ref(null)
-const deleteLoadingVersionId = ref(null)
-const versionPublishRecordDialogVisible = ref(false)
-const activeVersionPublishRecords = ref([])
-const activeVersionForRecords = ref(null)
 const schedulerSwitchDialogVisible = ref(false)
 const schedulerSwitchSaving = ref(false)
 const schedulerSwitchForm = reactive({
@@ -1134,11 +1130,24 @@ const {
   stepVersionCompare,
 } = useVersionCompare({ workflow, versionList, changeMode })
 
-const versionPublishRecordDialogTitle = computed(() => {
-  const versionNo = activeVersionForRecords.value?.versionNo
-  return Number.isFinite(Number(versionNo))
-    ? `版本 ${versionNo} 发布记录`
-    : '发布记录'
+const {
+  rollbackLoadingVersionId,
+  deleteLoadingVersionId,
+  versionPublishRecordDialogVisible,
+  activeVersionPublishRecords,
+  activeVersionForRecords,
+  versionPublishRecordDialogTitle,
+  rollbackToVersion,
+  deleteVersion,
+  openVersionPublishRecords,
+} = useVersionMutations({
+  workflow,
+  versionById,
+  publishRecordsByVersionId,
+  getRollbackDisabledReason: (...args) => getRollbackDisabledReason(...args),
+  getVersionDeleteDisabledReason: (...args) => getVersionDeleteDisabledReason(...args),
+  loadWorkflowDetail: (...args) => loadWorkflowDetail(...args),
+  backToPublishRecords: (...args) => backToPublishRecords(...args),
 })
 
 // Inline edit methods
@@ -1363,60 +1372,6 @@ const isVersionSelectable = (row) => {
   return selectedHistoryVersionIds.value.includes(versionId)
 }
 
-const openVersionPublishRecords = (row) => {
-  activeVersionForRecords.value = row || null
-  const versionId = Number(row?.id)
-  if (Number.isFinite(versionId)) {
-    activeVersionPublishRecords.value = publishRecordsByVersionId.value[versionId] || []
-  } else {
-    activeVersionPublishRecords.value = []
-  }
-  versionPublishRecordDialogVisible.value = true
-}
-
-const rollbackToVersion = async (versionId) => {
-  const wf = workflow.value?.workflow
-  const normalizedVersionId = Number(versionId)
-  if (!wf?.id || !Number.isFinite(normalizedVersionId)) {
-    return
-  }
-  const version = versionById.value[normalizedVersionId]
-  const rollbackDisabledReason = getRollbackDisabledReason(version)
-  if (rollbackDisabledReason) {
-    ElMessage.warning(rollbackDisabledReason)
-    return
-  }
-  const label = version ? `版本 ${version.versionNo}` : `#${versionId}`
-  try {
-    await ElMessageBox.confirm(
-      `确认恢复到${label}吗？恢复后会生成一个新版本。`,
-      '确认恢复',
-      {
-        type: 'warning',
-        confirmButtonText: '确认恢复',
-        cancelButtonText: '取消'
-      }
-    )
-  } catch {
-    return
-  }
-
-  rollbackLoadingVersionId.value = normalizedVersionId
-  try {
-    const response = await workflowApi.rollbackVersion(wf.id, normalizedVersionId, {
-      operator: 'portal-ui'
-    })
-    ElMessage.success(`恢复成功，已生成版本 v${response.newVersionNo}`)
-    backToPublishRecords()
-    await loadWorkflowDetail()
-  } catch (error) {
-    console.error('恢复版本失败', error)
-    ElMessage.error(error.message || '恢复版本失败')
-  } finally {
-    rollbackLoadingVersionId.value = null
-  }
-}
-
 const getRollbackDisabledReason = (row) =>
   rollbackDisabledReason(row, workflow.value?.workflow?.currentVersionId)
 
@@ -1426,45 +1381,6 @@ const getVersionDeleteDisabledReason = (row) =>
     workflow.value?.workflow?.currentVersionId,
     lastSuccessfulPublishedVersionId.value
   )
-
-const deleteVersion = async (row) => {
-  const wf = workflow.value?.workflow
-  const versionId = Number(row?.id)
-  if (!wf?.id || !Number.isFinite(versionId)) {
-    return
-  }
-  const disabledReason = getVersionDeleteDisabledReason(row)
-  if (disabledReason) {
-    ElMessage.warning(disabledReason)
-    return
-  }
-  const label = row?.versionNo ? `版本 ${row.versionNo}` : `#${versionId}`
-  try {
-    await ElMessageBox.confirm(
-      `确认删除${label}吗？删除后不可恢复。`,
-      '确认删除版本',
-      {
-        type: 'warning',
-        confirmButtonText: '确认删除',
-        cancelButtonText: '取消'
-      }
-    )
-  } catch {
-    return
-  }
-
-  deleteLoadingVersionId.value = versionId
-  try {
-    await workflowApi.deleteVersion(wf.id, versionId)
-    ElMessage.success('版本删除成功')
-    await loadWorkflowDetail()
-  } catch (error) {
-    console.error('删除版本失败', error)
-    ElMessage.error(error.message || '删除版本失败')
-  } finally {
-    deleteLoadingVersionId.value = null
-  }
-}
 
 const buildDolphinWorkflowUrl = (workflow) => {
   if (!dolphinWebuiUrl.value || !workflow?.projectCode || !workflow?.workflowCode) {
